#
Update Makefile and README for clarity and usability

Updated the `DISC_LABEL` and `BDJSDK_HOME` variables in the
Makefile to use correct values and paths. Simplified and
clarified installation steps in the README, including
removing user-specific prompts, correcting paths, and
renaming "Installation Steps" to "Building Steps." Added
a new contributor to the credits section for BD-JB-1250.